split loadable search results section out from main app
