+++ frontend/source/js/data-explorer/components/app/loadable-search-results.jsx
@@ -1,0 +1,6 @@
+import LoadableWrapper from '../loadable-wrapper';
+
+const LoadableSearchResults = LoadableWrapper(
+  () => import('./search-results'));
+
+export default LoadableSearchResults;
